Workspace error
parcel was not running, with an error of 'Segmentation fault: 11'

So, I ran npm update, re-installed eslint and prettier in VS Code, then installed prettier and eslint locally in the project folder

But now, it is only using double quotes in JSX, whereas before it was set to single quotes. The settings for the machine AND workspace both specify single quotes.

But I am glad that pacel will run and the segmentation error is no longer happening.

So, all the files were updated with double quotes.