Profile-View Fetch Bug Fixed
I did not realize, but the way I wrote the code for profile view caused the page to request server data hundreds of times a minute.

There were no "errors" thrown, because it was valid.

This happened for a few weeks now. Heroku is charging me $2.

Anyway, it's fixed now.