Updating minor changes
For movie-view, the close button has been changed to go back (-1) in history:

If you look at movie-view from the favorites within user profile, now it will go back to the user profile.

Formatted birthdate to Intl so that it will populate in the update-user form.

Put lonely containers within cards to achieve more uniform style

Added a glowing triangle as a logo in nav bar

Changed bg-gradient to reflect the variable colors used, not the theme default.

** Minor layout adjustments **

If a movie only had one similar movie, it would display in a strange way. Now it will render as expected.