RE-FORMAT SIGN UP AND PROFILE
I re-wrote the code to use less variables and possibly run better.

Combined update-user into profile-view, as it felt complicated to fix update-user while it was a separate component.

similar-movies:

if there was only one movie, it would display the card as SQUEEZED, so I added a min-width property there.